:recycle: Refactor firebase functions
