+++ functions/.gitignore
@@ -2,4 +2,4 @@ node_modules/
 
 firebase-*
 
-secrets/*
+config/*
