+++ functions/index.js
@@ -1,40 +1,14 @@
 const functions = require('firebase-functions');
 const app = require('express')();
 
-const admin = require('firebase-admin');
+// const cors = require('cors');
+// app.use(cors());
 
-admin.initializeApp();
+const Projects = require('./controllers/projects');
+const Users = require('./controllers/users');
 
-const db = admin.firestore();
-
-app.get('/projects', (req, res) => {
-  db.collection('projects')
-    .get()
-    .then(data => {
-      const projects = [];
-      data.forEach(doc => {
-        projects.push(doc.data());
-      });
-      res.json(projects);
-    })
-    .catch(err => console.log(err));
-});
-
-app.post('projects', (req, res) => {
-  const newPrject = {
-    name: req.body.name,
-    createdAt: admin.firestore.Timestamp.fromDate(new Date()),
-  };
-
-  db.collection('projects')
-    .add(newPrject)
-    .then(doc => {
-      res.json({ messages: `document ${doc.id} was created successfully` });
-    })
-    .catch(err => {
-      res.status(500).json({ error: `something went wrong` });
-      console.log(err);
-    });
-});
+app.get('/projects', Projects.index);
+app.post('/projects', Projects.create);
+app.post('/signup', Users.create);
 
 exports.api = functions.region('europe-west1').https.onRequest(app);
