+++ functions/package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "express": "^4.17.1",
+    "firebase": "^7.8.2",
     "firebase-admin": "^8.6.0",
     "firebase-functions": "^3.3.0"
   },
